Распределяю модели по подпапкам
