+++ frontend/controllers/TasksController.php
@@ -6,7 +6,7 @@ namespace frontend\controllers;
 
 use Yii;
 use yii\web\Controller;
-use app\models\Tasks;
+use app\models\tasks\Tasks;
 
 class TasksController extends Controller
 {
