+++ frontend/controllers/UsersController.php
@@ -6,7 +6,7 @@ namespace frontend\controllers;
 
 use Yii;
 use yii\web\Controller;
-use app\models\Users;
+use app\models\users\Users;
 
 class UsersController extends Controller
 {
