--- frontend/models/Categories.php
@@ -1,93 +1,0 @@
-<?php
-
-namespace app\models;
-
-use Yii;
-
-/**
- * This is the model class for table "categories".
- *
- * @property int $id
- * @property string $name
- * @property string $icon
- * @property string|null $profession
- *
- * @property Tasks[] $tasks
- * @property UserCategory[] $userCategories
- */
-class Categories extends \yii\db\ActiveRecord
-{
-    /**
-     * {@inheritdoc}
-     */
-    public static function tableName()
-    {
-        return 'categories';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function rules()
-    {
-        return [
-            [['name', 'icon'], 'required'],
-            [['name', 'icon', 'profession'], 'string', 'max' => 255],
-            [['name'], 'unique'],
-            [['icon'], 'unique'],
-            [['profession'], 'unique'],
-        ];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function attributeLabels()
-    {
-        return [
-            'id' => 'ID',
-            'name' => 'Name',
-            'icon' => 'Icon',
-            'profession' => 'Profession',
-        ];
-    }
-
-    /**
-     * Gets query for [[Tasks]].
-     *
-     * @return \yii\db\ActiveQuery|yii\db\ActiveQuery
-     */
-    public function getTasks()
-    {
-        return $this->hasMany(Tasks::class, ['category_id' => 'id']);
-    }
-
-    /**
-     * Gets query for [[UserCategories]].
-     *
-     * @return \yii\db\ActiveQuery|yii\db\ActiveQuery
-     */
-    public function getUserCategory()
-    {
-        return $this->hasMany(UserCategory::class, ['category_id' => 'id']);
-    }
-
-    /**
-     * Gets query for [[Users]].
-     *
-     * @return \yii\db\ActiveQuery|UsersQuery
-     */
-    public function getUsers()
-    {
-        return $this->hasMany(Users::class, ['id' => 'user_id'])->viaTable('user_category', ['categories_id' => 'id']);
-    }
-
-    /**
-     * {@inheritdoc}
-     * @return CategoriesQuery the active query used by this AR class.
-     */
-    public static function find()
-    {
-        return new CategoriesQuery(get_called_class());
-    }
-}
