--- frontend/models/CategoriesQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[Categories]].
- *
- * @see Categories
- */
-class CategoriesQuery extends \yii\db\ActiveQuery
-{
-}
