--- frontend/models/CitiesQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[Cities]].
- *
- * @see Cities
- */
-class CitiesQuery extends \yii\db\ActiveQuery
-{
-}
