--- frontend/models/FavouritesQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[Favourites]].
- *
- * @see Favourites
- */
-class FavouritesQuery extends \yii\db\ActiveQuery
-{
-}
