--- frontend/models/FileTaskQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[FileTask]].
- *
- * @see FileTask
- */
-class FileTaskQuery extends \yii\db\ActiveQuery
-{
-}
