--- frontend/models/MessagesQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[Messages]].
- *
- * @see Messages
- */
-class MessagesQuery extends \yii\db\ActiveQuery
-{
-}
