--- frontend/models/NotificationsQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[Notifications]].
- *
- * @see Notifications
- */
-class NotificationsQuery extends \yii\db\ActiveQuery
-{
-}
