--- frontend/models/OpinionsQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[Opinions]].
- *
- * @see Opinions
- */
-class OpinionsQuery extends \yii\db\ActiveQuery
-{
-}
