--- frontend/models/PortfolioPhotoQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[PortfolioPhoto]].
- *
- * @see PortfolioPhoto
- */
-class PortfolioPhotoQuery extends \yii\db\ActiveQuery
-{
-}
