--- frontend/models/RepliesQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[Replies]].
- *
- * @see Replies
- */
-class RepliesQuery extends \yii\db\ActiveQuery
-{
-}
