--- frontend/models/StatusTaskQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[StatusTask]].
- *
- * @see StatusTask
- */
-class StatusTaskQuery extends \yii\db\ActiveQuery
-{
-}
