--- frontend/models/TasksQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[Tasks]].
- *
- * @see Tasks
- */
-class TasksQuery extends \yii\db\ActiveQuery
-{
-}
