--- frontend/models/UserCategoryQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[UserCategory]].
- *
- * @see UserCategory
- */
-class UserCategoryQuery extends \yii\db\ActiveQuery
-{
-}
