--- frontend/models/UserRoleQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[UserRole]].
- *
- * @see UserRole
- */
-class UserRoleQuery extends \yii\db\ActiveQuery
-{
-}
