--- frontend/models/UsersQuery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace app\models;
-
-/**
- * This is the ActiveQuery class for [[Users]].
- *
- * @see Users
- */
-class UsersQuery extends \yii\db\ActiveQuery
-{
-}
