+++ frontend/models/categories/CategoriesQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\categories;
+
+/**
+ * This is the ActiveQuery class for [[Categories]].
+ *
+ * @see Categories
+ */
+class CategoriesQuery extends \yii\db\ActiveQuery
+{
+}
