+++ frontend/models/cities/Cities.php
@@ -1,0 +1,84 @@
+<?php
+
+namespace app\models\cities;
+
+use Yii;
+use app\models\{
+    tasks\Tasks,
+    users\Users
+};
+
+/**
+ * This is the model class for table "cities".
+ *
+ * @property int $id
+ * @property string $city
+ * @property string $latitude
+ * @property string $longitude
+ *
+ * @property Tasks[] $tasks
+ * @property Users[] $users
+ */
+class Cities extends \yii\db\ActiveRecord
+{
+    /**
+     * {@inheritdoc}
+     */
+    public static function tableName()
+    {
+        return 'cities';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function rules()
+    {
+        return [
+            [['city', 'latitude', 'longitude'], 'required'],
+            [['city', 'latitude', 'longitude'], 'string', 'max' => 255],
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function attributeLabels()
+    {
+        return [
+            'id' => 'ID',
+            'city' => 'City',
+            'latitude' => 'Latitude',
+            'longitude' => 'Longitude',
+        ];
+    }
+
+    /**
+     * Gets query for [[Tasks]].
+     *
+     * @return \yii\db\ActiveQuery|yii\db\ActiveQuery
+     */
+    public function getTasks()
+    {
+        return $this->hasMany(Tasks::class, ['city_id' => 'id']);
+    }
+
+    /**
+     * Gets query for [[Users]].
+     *
+     * @return \yii\db\ActiveQuery|yii\db\ActiveQuery
+     */
+    public function getUsers()
+    {
+        return $this->hasMany(Users::class, ['city_id' => 'id']);
+    }
+
+    /**
+     * {@inheritdoc}
+     * @return CitiesQuery the active query used by this AR class.
+     */
+    public static function find()
+    {
+        return new CitiesQuery(get_called_class());
+    }
+}
