+++ frontend/models/cities/CitiesQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\cities;
+
+/**
+ * This is the ActiveQuery class for [[Cities]].
+ *
+ * @see Cities
+ */
+class CitiesQuery extends \yii\db\ActiveQuery
+{
+}
