+++ frontend/models/messages/MessagesQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\messages;
+
+/**
+ * This is the ActiveQuery class for [[Messages]].
+ *
+ * @see Messages
+ */
+class MessagesQuery extends \yii\db\ActiveQuery
+{
+}
