+++ frontend/models/notifications/NotificationsQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\notifications;
+
+/**
+ * This is the ActiveQuery class for [[Notifications]].
+ *
+ * @see Notifications
+ */
+class NotificationsQuery extends \yii\db\ActiveQuery
+{
+}
