+++ frontend/models/opinions/OpinionsQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\opinions;
+
+/**
+ * This is the ActiveQuery class for [[Opinions]].
+ *
+ * @see Opinions
+ */
+class OpinionsQuery extends \yii\db\ActiveQuery
+{
+}
