+++ frontend/models/replies/RepliesQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\replies;
+
+/**
+ * This is the ActiveQuery class for [[Replies]].
+ *
+ * @see Replies
+ */
+class RepliesQuery extends \yii\db\ActiveQuery
+{
+}
