+++ frontend/models/tasks/FileTaskQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\tasks;
+
+/**
+ * This is the ActiveQuery class for [[FileTask]].
+ *
+ * @see FileTask
+ */
+class FileTaskQuery extends \yii\db\ActiveQuery
+{
+}
