+++ frontend/models/tasks/StatusTaskQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\tasks;
+
+/**
+ * This is the ActiveQuery class for [[StatusTask]].
+ *
+ * @see StatusTask
+ */
+class StatusTaskQuery extends \yii\db\ActiveQuery
+{
+}
