+++ frontend/models/tasks/TasksQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\tasks;
+
+/**
+ * This is the ActiveQuery class for [[Tasks]].
+ *
+ * @see Tasks
+ */
+class TasksQuery extends \yii\db\ActiveQuery
+{
+}
