+++ frontend/models/users/FavouritesQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\users;
+
+/**
+ * This is the ActiveQuery class for [[Favourites]].
+ *
+ * @see Favourites
+ */
+class FavouritesQuery extends \yii\db\ActiveQuery
+{
+}
