+++ frontend/models/users/PortfolioPhotoQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\users;
+
+/**
+ * This is the ActiveQuery class for [[PortfolioPhoto]].
+ *
+ * @see PortfolioPhoto
+ */
+class PortfolioPhotoQuery extends \yii\db\ActiveQuery
+{
+}
