+++ frontend/models/users/UserCategoryQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\users;
+
+/**
+ * This is the ActiveQuery class for [[UserCategory]].
+ *
+ * @see UserCategory
+ */
+class UserCategoryQuery extends \yii\db\ActiveQuery
+{
+}
