+++ frontend/models/users/UserRoleQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\users;
+
+/**
+ * This is the ActiveQuery class for [[UserRole]].
+ *
+ * @see UserRole
+ */
+class UserRoleQuery extends \yii\db\ActiveQuery
+{
+}
