+++ frontend/models/users/UsersQuery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace app\models\users;
+
+/**
+ * This is the ActiveQuery class for [[Users]].
+ *
+ * @see Users
+ */
+class UsersQuery extends \yii\db\ActiveQuery
+{
+}
